Put `pub` function in top level `Polyn` module
